# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatnow API",
        "description": "Whatnow API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.whatnow.jazusoft.com/v1",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/org": {
            "get": {
                "tags": [
                    "Organisation"
                ],
                "summary": "Get all organisations (public)",
                "operationId": "04f86edf4b63c5b0f297e462633924b4",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/org/{code}": {
            "get": {
                "tags": [
                    "Organisation"
                ],
                "summary": "Get organisation by country code (public)",
                "operationId": "b4a9c9d91b102ab0d0709eae9104db61",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Country code of the organisation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Organisation"
                ],
                "summary": "Update an organisation by its country code",
                "description": "Updates the details of an organisation based on the provided country code.",
                "operationId": "OrganisationController@putById",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Country code of the organisation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data to update the organisation",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "https://example.com",
                                        "nullable": true
                                    },
                                    "translations": {
                                        "type": "object",
                                        "additionalProperties": {
                                            "properties": {
                                                "languageCode": {
                                                    "type": "string",
                                                    "example": "en"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "maxLength": 255,
                                                    "example": "International Federation of Red Cross and Red Crescent Societies"
                                                },
                                                "attributionMessage": {
                                                    "type": "string",
                                                    "maxLength": 2048,
                                                    "example": "attributionMessage example"
                                                },
                                                "contributors": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 111
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "This is an example string with exactly one hundred characters, including spaces and symbols"
                                                            },
                                                            "logo": {
                                                                "type": "string",
                                                                "example": "https://whatnowimages.blob.core.windows.net/images/LOGO"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "published": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
...
        "/whatnow/{id}": {
            "get": {
                "tags": [
                    "Whatnow"
                ],
                "summary": "Obtiene un recurso publicado por ID (public)",
                "description": "Retorna los detalles de un recurso publicado basado en el ID proporcionado.",
                "operationId": "getPublishedById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del recurso publicado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Whatnow"
                ],
                "summary": "Update a WhatNow entity by ID",
                "operationId": "putById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the WhatNow entity to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "countryCode",
                                    "eventType",
                                    "translations"
                                ],
                                "properties": {
                                    "countryCode": {
                                        "description": "Country code (3 characters)",
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "eventType": {
                                        "description": "Type of event (max 50 characters)",
                                        "type": "string",
                                        "example": "Flood"
                                    },
                                    "regionName": {
                                        "description": "Name of the subnational (optional)",
                                        "type": "string",
                                        "example": "North Region"
                                    },
                                    "translations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "webUrl": {
                                                    "description": "Web URL for the translation (optional)",
                                                    "type": "string",
                                                    "example": "https://example.com"
                                                },
                                                "lang": {
                                                    "description": "Language code (2 characters)",
                                                    "type": "string",
                                                    "example": "en"
                                                },
                                                "title": {
                                                    "description": "Title in the specified language",
                                                    "type": "string",
                                                    "example": "Flood Alert"
                                                },
                                                "description": {
                                                    "description": "Description in the specified language",
                                                    "type": "string",
                                                    "example": "Description of the event"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            },
            "delete": {
                "tags": [
                    "Whatnow"
                ],
                "summary": "Delete a WhatNow entity by ID",
                "operationId": "deleteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the WhatNow entity to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key"
            }
        }
    },
    "security": [
        {
            "BasicAuth": []
        }
    ]
}